rework test into integration, since CH instance is needed
